win32 patches
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@3297 d0c010d8-bb82-477e-a075-112140230451

+++ 3rdparty/cache2gtiff/main.cpp
@@ -299,7 +299,7 @@ int main(int argc, char ** argv)
 
     if(isTMS)
     {
-        level = log(level)/log(2) + 1;
+        level = log((double)level)/log(2.0) + 1;
         return exportTMS(level, lon1, lat1, lon2, lat2, infile, outfile, diskCache);
     }
     else
